- Fix submodule import after refactor - Fix test for run_model interface
