+++ neural_lifetimes/models/nets/__init__.py
@@ -2,7 +2,7 @@ from .custom_par import DeepEchoModel, PAREdit, PARModel, PARNet
 from .custom_tvae import DataTransformer, Decoder, TVAESynthesizer
 from .embedder import CombinedEmbedder
 from .encoder_decoder import VariationalEncoderDecoder
-from .neural_lifetimes import EventEncoder
+from .event_model import EventEncoder
 
 __all__ = [
     DeepEchoModel,
