+++ tests/test_interface/test_run_model.py
@@ -123,8 +123,8 @@ class TestRunModel:
             trainer_kwargs={"accumulate_grad_batches": 2, "callbacks": []},
         )
         assert (
-            len(trainer.callbacks) <= 2
-        )  # progressbar callback is always given and modelcheckpoints are on by default
+            len(trainer.callbacks) <= 4
+        )  # progressbar, model summary and model checkpoints are turned on by default. Gradient Accummulation also enables => max 4 callbacks expected
         assert trainer.accumulate_grad_batches == 2
 
     def test_custom_logger_class(
